[flink] Refactor update data fields process function (#1822)
